PhotosTableViewController connected, but no predicates set for recents.
I've set up the PhotosTableViewController in the app delegate to be
connected to the most recent tab, but it is currently displaying
everything in title order.   Next, I'll try to configure the predicates
and sort descriptors to only get the most recent (within the last 48
hours) and sort them by lastViewed date.