Working on Favorites, fixed recents
Added attribute to the place model for hasFavorites, implemented method
on Photo to toggle this, and reset the button target/selector for
favorites. Also fixed the issue where recents weren't showing up if it
wasn't the first place you went. I did this by removing the cache - I
don't know - may not be the right thing to do, but it seems to have
solved my issue.