Favorites in place
Favorites are working. I'm not caching the photos for favorites yet,
that will be next. And there appears to be an issue with photos that do
not have a name. I display "Unkown" in the flickr controller for that,
but that "Unknown" doesn't get persisted, so it doesn't display
elsewhere at the moment.